Moving the WebAppConfiguration class.
