+++ gui-web/src/main/java/de/ollie/counter/ws/gui/vaadin/counters/CountersView.java
@@ -19,10 +19,10 @@ import de.ollie.counter.ws.core.service.CounterService;
 import de.ollie.counter.ws.core.service.UserService;
 import de.ollie.counter.ws.core.service.localization.ResourceManager;
 import de.ollie.counter.ws.gui.SessionData;
-import de.ollie.counter.ws.gui.WebAppConfiguration;
 import de.ollie.counter.ws.gui.vaadin.GUIConfiguration;
 import de.ollie.counter.ws.gui.vaadin.MainMenuView;
 import de.ollie.counter.ws.gui.vaadin.UserAuthorizationChecker;
+import de.ollie.counter.ws.gui.vaadin.WebAppConfiguration;
 import de.ollie.counter.ws.gui.vaadin.component.ButtonFactory;
 import de.ollie.counter.ws.gui.vaadin.component.HeaderLayout;
 import de.ollie.counter.ws.gui.vaadin.component.HeaderLayout.HeaderLayoutMode;
