+++ service/src/main/resources/dm/counter-ws.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<Diagram application="Archimedes" application-version="2.33.17" diagram-version="2.0">
+<Diagram application="Archimedes" application-version="2.34.0" diagram-version="2.0">
     <AdditionalSQLStatements postChanging="" postReducing="" preChanging="" preExtending="" />
     <Colors>
         <Color b="255" g="0" name="blau" r="0" />
@@ -49,6 +49,7 @@
     <Options>
         <Option name="ALTERNATE_GUI_VAADIN_MODULE_PREFIX" parameter="gui-web" />
         <Option name="ALTERNATE_MODULE_PREFIX" parameter="" />
+        <Option name="CUBE_APPLICATION" parameter="" />
         <Option name="MODULE_MODE" parameter="" />
         <Option name="REFERENCE_MODE" parameter="OBJECT" />
         <Option name="SCHEMA_NAME" parameter="counterws" />
@@ -121,6 +122,7 @@
             <Column canBeReferenced="false" comment="" deprecated="false" domain="Varchar255" editorAlterInBatchView="false" editorDisabled="false" editorLabelText="" editorMaxLength="0" editorMember="false" editorMnemonic="" editorPosition="0" editorReferenceWeight="NONE" editorResourceId="" editorToolTipText="" encrypted="false" global="false" globalId="false" hasIndex="false" hideReference="false" history="" indexSearchMember="false" individualDefaultValue="" lastModificationField="false" listItemField="false" name="TAG" notNull="false" panelNumber="0" parameter="" primaryKey="false" removedStateField="false" required="false" sequenceForKeyGeneration="" suppressForeignKeyConstraints="false" technicalField="false" transient="false" unique="false" />
             <Options>
                 <Option name="NO_GENERATION" parameter="" />
+                <Option name="SUPPRESS_NO_PK_WARNING" parameter="" />
             </Options>
             <Panel panelClass="" panelNumber="0" tabMnemonic="1" tabTitle="data" tabToolTipText="Hier k&amp;ouml;nnen Sie die Daten des Objekt warten" />
             <View name="Main" x="375" y="875" />
